+++ src/cpu/jit_avx512_common_1x1_conv_kernel.cpp
@@ -295,7 +295,7 @@ void jit_avx512_common_1x1_conv_kernel::reduce_loop(int load_loop_blk,
             if (jcp.relu_negative_slope == 0) {
                 zmm_relu_ns = zmm_zero;
             } else {
-                mov(imm_addr64, float2int(jcp.relu_negative_slope));
+                mov(imm_addr64, float2int((float)jcp.relu_negative_slope));
                 vmovq(xmm_relu_ns, imm_addr64);
                 vbroadcastss(zmm_relu_ns, xmm_relu_ns);
             }
